#132 Fix missing 2020 Tokyo Olympics holiday relocations + tests

+++ holiday-calendar-apac/src/test/java/org/holiday/calendar/impl/HolidayCalendarServiceJPExtendedTest.java
@@ -614,7 +614,7 @@ public class HolidayCalendarServiceJPExtendedTest {
                 .map(HolidayDate::getDate)
                 .distinct()
                 .count();
-        assertEquals(distinctCount, (long) holidays.size(),
+        assertEquals(distinctCount, holidays.size(),
                      year + ": holiday list must not contain duplicate observed dates");
     }
