+++ holiday-calendar-apac/src/test/java/org/holiday/calendar/impl/HolidayCalendarServiceJPTest.java
@@ -293,7 +293,7 @@ public class HolidayCalendarServiceJPTest {
     public void testJP_NoDateDuplicatesIn2021() {
         List<HolidayDate> holidays = service.getHolidayCalendar().calculate(2021);
         long uniqueDates = holidays.stream().map(HolidayDate::date).distinct().count();
-        assertEquals(uniqueDates, (long) holidays.size(),
+        assertEquals(uniqueDates, holidays.size(),
                 "2021: no two holidays should share a date");
     }
